Moved VersionedRunner to its own module file.
