+++ crates/bld_runner/src/sync/mod.rs
@@ -1,2 +1,3 @@
 pub mod builder;
-pub mod runner;
+mod runner;
+pub mod versioned;
